requirements.txt
flask
tensorflow
numpy
Pillow
streamlit
requests
Pillow
mediapipe
opencv-python-headless
pandas
plotly